pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "arduino-led-control"
version = "0.1.0"
description = "Arduino LED control module with compiled firmware in wheel"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Thomas Dietel", email = "tom@dietel.net"}
]
keywords = ["arduino", "serial", "led", "control"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
dependencies = [
    "pyserial>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/tdietel/arduino-led-control"
Documentation = "https://github.com/tdietel/arduino-led-control#readme"
Repository = "https://github.com/tdietel/arduino-led-control.git"
Issues = "https://github.com/tdietel/arduino-led-control/issues"

[tool.setuptools]
packages = ["arduino_led_control"]
include-package-data = true

[tool.setuptools.package-data]
arduino_led_control = ["firmware/*.hex", "firmware/*.bin"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312', 'py313']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
